requirements.txt
ffmpeg-python
librosa
mutagen
fastapi
uvicorn
pywebview
